requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.3.1
torchvision==0.18.1
pillow==10.4.0
numpy==1.26.4
